Rename halo2_pse exponentiation to exponentiate
