+++ halo2_pse/src/circuits/mod.rs
@@ -1,4 +1,4 @@
-pub mod exponentiation;
+pub mod exponentiate;
 pub mod is_zero;
 pub mod sha256;
 pub mod utils;
